requirements/test.txt
PyYAML>=3.10
nose
coverage
